test: fixes and remove unnecesary tests
