+++ crates/pop-cli/src/commands/up/rollup.rs
@@ -552,7 +552,6 @@ mod tests {
 	const MOCK_PROXIED_ADDRESS: &str = "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty";
 	const MOCK_PROXY_ADDRESS_ID: &str = "Id(13czcAAt6xgLwZ8k6ZpkrRL5V2pjKEui3v9gHAN9PoxYZDbf)";
 	const POLKADOT_NETWORK_URL: &str = "wss://polkadot-rpc.publicnode.com";
-	const POP_NETWORK_TESTNET_URL: &str = "wss://rpc1.paseo.popnetwork.xyz";
 
 	#[tokio::test]
 	async fn prepare_for_deployment_works() -> Result<()> {
@@ -651,9 +650,7 @@ mod tests {
 		Ok(())
 	}
 
-	// TODO: Slow -> 30s
 	#[tokio::test]
-	#[ignore]
 	async fn prepare_for_registration_works() -> Result<()> {
 		let node = TestNode::spawn().await?;
 		let node_url = node.ws_url();
@@ -742,9 +739,10 @@ mod tests {
 		cli.verify()
 	}
 
-	// TODO: Integration test. Querying Polkadot
 	#[tokio::test]
 	async fn register_fails_wrong_chain() -> Result<()> {
+		let node = TestNode::spawn().await?;
+		let node_url = node.ws_url();
 		let mut cli = MockCli::new()
             .expect_intro("Deploy a rollup")
             .expect_select(
@@ -764,7 +762,7 @@ mod tests {
                 DeploymentProvider::VARIANTS.len(), // Register
                 None,
             )
-            .expect_info(format!("You will need to sign a transaction to register on {}, using the `Registrar::register` function.", Url::parse(POP_NETWORK_TESTNET_URL)?.as_str()))
+            .expect_info(format!("You will need to sign a transaction to register on {}, using the `Registrar::register` function.", Url::parse(node_url)?.as_str()))
             .expect_outro_cancel(format!("Failed to find the pallet Registrar\n{}", style(format!(
 				"Retry registration without reserve or rebuilding the chain specs using: {}", style("`pop up --id 2000 --skip-registration`").bold()
 			)).black()
@@ -774,7 +772,7 @@ mod tests {
 			id: Some(2000),
 			genesis_state: Some(genesis_state.clone()),
 			genesis_code: Some(genesis_code.clone()),
-			relay_chain_url: Some(Url::parse(POP_NETWORK_TESTNET_URL)?),
+			relay_chain_url: Some(Url::parse(node_url)?),
 			path: None,
 			proxied_address: None,
 			..Default::default()
@@ -831,9 +829,10 @@ mod tests {
 		Ok(())
 	}
 
-	// TODO: Integration test. Querying Polkadot
 	#[tokio::test]
 	async fn reserve_id_fails_wrong_chain() -> Result<()> {
+		let node = TestNode::spawn().await?;
+		let node_url = node.ws_url();
 		let mut cli = MockCli::new()
             .expect_intro("Deploy a rollup")
             .expect_select(
@@ -853,14 +852,14 @@ mod tests {
                 DeploymentProvider::VARIANTS.len(), // Register
                 None,
             )
-            .expect_info(format!("You will need to sign a transaction to reserve an ID on {} using the `Registrar::reserve` function.", Url::parse(POP_NETWORK_TESTNET_URL)?.as_str()))
+            .expect_info(format!("You will need to sign a transaction to reserve an ID on {} using the `Registrar::reserve` function.", Url::parse(node_url)?.as_str()))
             .expect_outro_cancel("Failed to find the pallet Registrar");
 		let (genesis_state, genesis_code) = create_temp_genesis_files()?;
 		UpCommand {
 			id: None,
 			genesis_state: Some(genesis_state.clone()),
 			genesis_code: Some(genesis_code.clone()),
-			relay_chain_url: Some(Url::parse(POP_NETWORK_TESTNET_URL)?),
+			relay_chain_url: Some(Url::parse(node_url)?),
 			path: None,
 			proxied_address: None,
 			..Default::default()
