+++ crates/pop-cli/src/common/try_runtime.rs
@@ -627,9 +627,7 @@ mod tests {
 		Ok(())
 	}
 
-	// TODO: Slow -> 45s
 	#[tokio::test]
-	#[ignore]
 	async fn update_live_state_works() -> anyhow::Result<()> {
 		let node = TestNode::spawn().await?;
 		let node_url = node.ws_url();
@@ -757,9 +755,7 @@ mod tests {
 		Ok(())
 	}
 
-	// TODO: Slow -> 45.9s
 	#[tokio::test]
-	#[ignore]
 	async fn guide_user_to_select_try_state_works() -> anyhow::Result<()> {
 		let node = TestNode::spawn().await?;
 		let node_url = node.ws_url();
